Add files via upload
Snakesandladders.ipynb is a jupyter notebook project file from an object oriented project to analyse the game of snakes and ladders. To begin with, the original snakes and ladders board is set up, and the probability of where each turn is ended is used to calculate interesting statistics such as the average game length. Some interesting plots are also produced. Later in the project, user inputs for a unique board are taken. The user defines the size of the board, as well as the number of snakes and the number of ladders on the board. The code takes these inputs, randomizes the locations of all the snakes and ladders and generates the unique transfer matrix for this board, which is the driving engine of the calculations done.